refactor caching metrics to use Prometheus counters and update Grafan…

+++ ArticleService/Model/ArticleMetrics.cs
@@ -2,8 +2,8 @@ namespace ArticleService.Model;
 
 public class ArticleMetrics
 {
-    public int Hits { get; set; }
-    public int Misses { get; set; }
+    public double Hits { get; set; }
+    public double Misses { get; set; }
     public List<int> ArticleIdsInCache { get; set; }
     public double CacheHitRatio { get; set; }
 }
