+++ CommentService/Models/CommentMetrics.cs
@@ -2,8 +2,8 @@ namespace CommentService.Models;
 
 public class CommentMetrics
 {
-    public int Hits { get; set; }
-    public int Misses { get; set; }
+    public double Hits { get; set; }
+    public double Misses { get; set; }
     public LinkedList<int> LruList { get; set; }
     public double CacheHitRatio { get; set; }
 }
